requirements.txt
black
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
slack_bolt
